+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/feature/registrables/hooks/base/HookExecutor.java
@@ -1,14 +1,14 @@
 package com.samares_engineering.omf.omf_core_framework.feature.registrables.hooks.base;
 
 
-import com.samares_engineering.omf.omf_core_framework.plugin.APlugin;
+import com.samares_engineering.omf.omf_core_framework.plugin.OMFPlugin;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public abstract class HookExecutor<H extends Hook> {
     List<H> hooksHolders;
-    private APlugin plugin;
+    private OMFPlugin plugin;
 
     public void addHook(H hook) {
         hooksHolders.add(hook);
@@ -32,12 +32,12 @@ public abstract class HookExecutor<H extends Hook> {
     }
 
 
-    public void init(APlugin aPlugin) {
-        this.plugin = aPlugin;
+    public void init(OMFPlugin OMFPlugin) {
+        this.plugin = OMFPlugin;
         this.hooksHolders = new ArrayList<>();
     }
 
-    public APlugin getPlugin() {
+    public OMFPlugin getPlugin() {
         return plugin;
     }
 }