+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/model_comparators/filters/ElementFilter.java
@@ -78,7 +78,6 @@ public class ElementFilter implements ModelComparatorFilter {
                 return false;
             default:
                 return true;
-
         }
     }
 }