+++ omf-test-framework/src/main/java/com/samares_engineering/omf/omf_test_framework/plugin/OMFTestPlugin.java
@@ -1,0 +1,17 @@
+package com.samares_engineering.omf.omf_test_framework.plugin;
+
+import com.samares_engineering.omf.omf_core_framework.feature.OMFFeature;
+import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.configurators.implementations.OMFBrowserConfigurator;
+import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.configurators.implementations.OMFDiagramConfigurator;
+import com.samares_engineering.omf.omf_core_framework.feature.registrables.actions.configurators.implementations.OMFMainMenuConfigurator;
+import com.samares_engineering.omf.omf_core_framework.listeners.IListenerManager;
+import com.samares_engineering.omf.omf_core_framework.listeners.listeners.ProjectListener;
+import com.samares_engineering.omf.omf_core_framework.plugin.OMFPlugin;
+import com.samares_engineering.omf.omf_core_framework.ui.environmentoptions.OMFPropertyOptionsGroup;
+import com.samares_engineering.omf.omf_core_framework.ui.projectoptions.FeatureProjectOptionsConfigurator;
+
+import java.util.List;
+
+public interface OMFTestPlugin extends OMFPlugin {
+
+}